+++ R/data.table.R
@@ -216,9 +216,6 @@ replace_dot_alias <- function(e) {
   }
 }
 
-# count of updated rows after `:=` stored in .Last.updated #1885
-.Last.updated <- NULL # onLoad init this to copy(NA_integer_)
-
 "[.data.table" <- function (x, i, j, by, keyby, with=TRUE, nomatch=getOption("datatable.nomatch"), mult="all", roll=FALSE, rollends=if (roll=="nearest") c(TRUE,TRUE) else if (roll>=0) c(FALSE,TRUE) else c(TRUE,FALSE), which=FALSE, .SDcols, verbose=getOption("datatable.verbose"), allow.cartesian=getOption("datatable.allow.cartesian"), drop=NULL, on=NULL)
 {
   # ..selfcount <<- ..selfcount+1  # in dev, we check no self calls, each of which doubles overhead, or could
@@ -1140,7 +1137,6 @@ replace_dot_alias <- function(e) {
           cols = as.integer(m)
           newnames=NULL
           if (identical(irows, integer())) {
-            .Call(Cassign, NULL, irows, NULL, NULL, NULL, FALSE, TRUE, get(".Last.updated", envir=topenv())) # call only to write 0 to .Last.updated
             # Empty integer() means no rows e.g. logical i with only FALSE and NA
             # got converted to empty integer() by the which() above
             # Short circuit and do-nothing since columns already exist. If some don't
@@ -1154,6 +1150,7 @@ replace_dot_alias <- function(e) {
               cat("No rows match i. No new columns to add so not evaluating RHS of :=\n")
               cat("Assigning to 0 row subset of",nrow(x),"rows\n")
             }
+            .Call(Cassign, x, irows, NULL, NULL, NULL, FALSE) # only purpose is to write 0 to .Last.updated
             .global$print = address(x)
             return(invisible(x))
           }
@@ -1385,7 +1382,7 @@ replace_dot_alias <- function(e) {
     }
     if (!is.null(lhs)) {
       # TODO?: use set() here now that it can add new columns. Then remove newnames and alloc logic above.
-      .Call(Cassign, x, irows, cols, newnames, jval, verbose, TRUE, get(".Last.updated", envir=topenv()))
+      .Call(Cassign,x,irows,cols,newnames,jval,verbose)
       return(suppPrint(x))
     }
     if ((is.call(jsub) && is.list(jval) && jsub[[1L]] != "get" && !is.object(jval)) || !missingby) {
@@ -2099,7 +2096,7 @@ tail.data.table <- function(x, n=6L, ...) {
     # search for one other .Call to assign in [.data.table to see how it differs
   }
   verbose=getOption("datatable.verbose")
-  x = .Call(Cassign,copy(x),i,cols,newnames,value,verbose,FALSE,NULL) # From 3.1.0, DF[2,"b"] = 7 no longer copies DF$a (so in this [<-.data.table method we need to copy)
+  x = .Call(Cassign,copy(x),i,cols,newnames,value,verbose) # From 3.1.0, DF[2,"b"] = 7 no longer copies DF$a (so in this [<-.data.table method we need to copy)
   alloc.col(x)  #  can maybe avoid this realloc, but this is (slow) [<- anyway, so just be safe.
   if (length(reinstatekey)) setkeyv(x,reinstatekey)
   invisible(x)
@@ -2642,7 +2639,7 @@ set <- function(x,i=NULL,j,value)  # low overhead, loopable
     .Call(Csetlistelt,l,1L,value)  # to avoid the copy by list() in R < 3.1.0
     value = l
   }
-  .Call(Cassign,x,i,j,NULL,value,FALSE,FALSE,NULL)   #  verbose=FALSE for speed to avoid getOption()  TO DO: somehow read getOption("datatable.verbose") from C level
+  .Call(Cassign,x,i,j,NULL,value,FALSE)   #  verbose=FALSE for speed to avoid getOption()  TO DO: somehow read getOption("datatable.verbose") from C level
   invisible(x)
 }